+++ newt/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2025-11-28
+
+### Added
+- Implement work by [adriaanConijn](https://github.com/adriaanConijn/hass-addon-newt) to include mTLS, DNS, and Log Level support.
+- Implement Docker socket support.
+
+### Changed
+- Newt version bumped to 1.6.0.
+
 ## [0.5.0] - 2025-06-20
 
 ### Changed